resolved some compiler warnings in unit tests

+++ src/io/remotesteerer/test/parallel_mpi_4/pipetest.h
@@ -31,7 +31,7 @@ public:
         TS_ASSERT_EQUALS(pipe.steeringRequests[0], "set heat 0.1 100 120 110");
         TS_ASSERT_EQUALS(pipe.steeringRequests[1], "set flow 6.9 100 120 110");
 
-        TS_ASSERT_EQUALS(pipe.retrieveSteeringRequests().size(), 2);
+        TS_ASSERT_EQUALS(pipe.retrieveSteeringRequests().size(), unsigned(2));
         TS_ASSERT_EQUALS(pipe.steeringRequests.size(), 0);
     }
 
@@ -47,7 +47,7 @@ public:
         pipe.addSteeringFeedback("node " + StringOps::itoa(mpiLayer.rank()) + " shutting down");
 
         pipe.sync();
-        int expectedSize = (mpiLayer.rank() == 0)? 9 : 0;
+        unsigned expectedSize = (mpiLayer.rank() == 0)? 9 : 0;
         TS_ASSERT_EQUALS(pipe.steeringFeedback.size(),           expectedSize);
         TS_ASSERT_EQUALS(pipe.copySteeringFeedback().size(),     expectedSize);
         TS_ASSERT_EQUALS(pipe.retrieveSteeringFeedback().size(), expectedSize);
